# feat: exclude `%EvtxFile%` profile duplication check when `-X` enabled