## What Changed

- Related #1158 
- Exclude `%EvtxFile%` profile duplication check when `-X` enabled
  - Because we don't want duplicate records to be output when the user analyzes including the backup Evtx.

## Evidence
### Environment
- OS: macOS montery version 13.1
- Hard: MacBook Air(M1, 2020) , Memory 8GB, Core 8

### Test
I checked hayabusa-sample-extx result, then result is as follows.
| Rev | Option | Total/Unique detections | File record count(`wc -l`) | File size |
| --- | ------ | ----------------------- | ----------------- | --------- |
| main | - | 32,266 / 568 | 32,267 | 32,297,127 |
| This PR | - | 32,266 / 568 | 32,267 | 32,297,127 |
| main | -p super-verbose | 32,266 / 568 | 32,267 | 39,229,313 |
| This PR  | -p super-verbose | 32,266 / 568 | 32,267 | 39,229,313  |
| main | -X | 32,050 / 568 | 32,051 | 32,073,306 |
| This PR  | -X | 32,050 / 568 | 32,051 | 32,073,306 |
| main | -X -p super-verbose | 32,266 / 568 | 32,267 | 39,229,313 |
| This PR  | -X -p super-verbose | 32,050 / 568 | 32,051 | 38,947,406 |

I would appreciate it if you could review🙏