@YamatoSecurity 2023-08-17
@fukusuket Thanks so much! I took some benchmarks. First on hayabusa-sample-evtx with duplicates of the evtx files inside the directory.

Main without X:

```
Saved file: test-x.csv (80.8 MB)

Elapsed time: 00:00:15.579
Rule Parse Processing Time: 00:00:02.621
Analysis Processing Time: 00:00:11.367
Output Processing Time: 00:00:01.590

Memory usage stats:
heap stats:     peak       total       freed     current        unit       count
  reserved:     2.0 GiB     2.0 GiB     0           2.0 GiB
 committed:     1.8 GiB     2.0 GiB    17.9 GiB   -15.9 GiB                          ok
...
   process: user: 73.135 s, system: 3.187 s, faults: 0, rss: 608.7 MiB, commit: 1.8 GiB
```

Main with -X:

```
Saved file: test-x.csv (80.8 MB)

Elapsed time: 00:00:15.779
Rule Parse Processing Time: 00:00:02.625
Analysis Processing Time: 00:00:11.489
Output Processing Time: 00:00:01.664

Memory usage stats:
heap stats:     peak       total       freed     current        unit       count
  reserved:     3.0 GiB     3.0 GiB     0           3.0 GiB
 committed:     2.8 GiB     3.0 GiB    16.8 GiB   -13.8 GiB                          ok
...
   process: user: 73.546 s, system: 3.290 s, faults: 0, rss: 612.9 MiB, commit: 2.8 GiB
```


This PR without X:
```
./target/release/hayabusa csv-timeline -u -D -P -E -d ../hayabusa-sample-evtx -o test-x.csv -p super-verbose -C --debug

Saved file: test-x.csv (80.8 MB)

Elapsed time: 00:00:16.691
Rule Parse Processing Time: 00:00:02.650
Analysis Processing Time: 00:00:12.258
Output Processing Time: 00:00:01.781

Memory usage stats:
heap stats:     peak       total       freed     current        unit       count
  reserved:     3.0 GiB     3.0 GiB     0           3.0 GiB
 committed:     2.8 GiB     3.0 GiB    17.5 GiB   -14.5 GiB                          ok
...
   process: user: 73.858 s, system: 3.376 s, faults: 0, rss: 639.2 MiB, commit: 2.8 GiB
```


This PR with -X:
```
./target/release/hayabusa csv-timeline -u -D -P -E -d ../hayabusa-sample-evtx -o test-x.csv -p super-verbose -C --debug -X

Saved file: test-x.csv (40.5 MB)

Elapsed time: 00:00:14.987
Rule Parse Processing Time: 00:00:02.629
Analysis Processing Time: 00:00:11.393
Output Processing Time: 00:00:00.964

Memory usage stats:
heap stats:     peak       total       freed     current        unit       count
  reserved:     3.0 GiB     3.0 GiB     0           3.0 GiB
 committed:     2.8 GiB     3.0 GiB    17.7 GiB   -14.7 GiB                          ok
...
   process: user: 71.720 s, system: 3.120 s, faults: 0, rss: 610.0 MiB, commit: 2.8 GiB
```

Now Hayabusa will ignore the duplicate files! Great!


On 608 CTF logs (92GB):
Main with -X:
```
./target/release/hayabusa csv-timeline -u -D -P -E -d ../608-logs -o test-x.csv -p super-verbose -C --debug -X
Saved file: test-x.csv (17.3 GB)

Elapsed time: 01:18:13.159
Errors were generated. Please check ./logs/errorlog-20230816_183417.log for details.

Rule Parse Processing Time: 00:00:03.570
Analysis Processing Time: 01:02:49.786
Output Processing Time: 00:15:19.792

Memory usage stats:
heap stats:     peak       total       freed     current        unit       count
  reserved:    80.0 GiB    82.0 GiB     2.0 GiB    80.0 GiB
 committed:     1.8 GiB    82.0 GiB  2606.3 GiB -2524.3 GiB                          ok
...
   process: user: 38147.694 s, system: 953.413 s, faults: 0, rss: 41.2 GiB, commit: 1.8 GiB
```

This PR:
```
./hayabusa-2.8.0-IgnoreEvtxFile  csv-timeline -u -D -P -E -d ../608-logs -o test-x-pr.csv -p super-verbose -C --debug -X

Saved file: test-x-pr.csv (17.3 GB)

Elapsed time: 01:23:27.774
Errors were generated. Please check ./logs/errorlog-20230816_222922.log for details.

Rule Parse Processing Time: 00:00:03.699
Analysis Processing Time: 01:06:01.804
Output Processing Time: 00:17:22.259

Memory usage stats:
heap stats:     peak       total       freed     current        unit       count
  reserved:    88.0 GiB    90.0 GiB     2.0 GiB    88.0 GiB
 committed:     1.8 GiB    90.0 GiB  2305.8 GiB -2215.8 GiB                          ok
...
   process: user: 38773.393 s, system: 1072.782 s, faults: 0, rss: 36.8 GiB, commit: 1.8 GiB
```

It seems that this PR is using more memory. 82GB -> 90GB.
Is it possible to implement it without `clone()`?
If it is difficult then no problem, we can implement it as is.