@fukusuket 2023-08-17
@YamatoSecurity 
Thank you so much for benchmarking! Yes, this clone was unnecessary...😅
I am sorry to bother you. Could you benchmark?